requirements.txt
asgiref==3.7.2
attrs==23.2.0
certifi==2024.6.2
charset-normalizer==3.3.2
Django==4.2.11
djangorestframework==3.15.2
exceptiongroup==1.2.1
h11==0.14.0
idna==3.7
outcome==1.3.0.post0
packaging==24.1
psycopg2==2.9.9
psycopg2-binary==2.9.9
PySocks==1.7.1
python-decouple==3.8
python-dotenv==1.0.1
requests==2.32.3
selenium==4.22.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.4.4
trio==0.25.1
trio-websocket==0.11.1
typing_extensions==4.10.0
urllib3==2.2.2
webdriver-manager==4.0.1
websocket-client==1.8.0
wsproto==1.2.0
